ref(server): More cleanup
